Add NumberOfChecks.py
Compute number of checks by getting a set of squares indicating squares with a piece that puts the opponent's king in check. Convert that set of squares to a list of boolean values and then sum the entries which are true. This gives the number of pieces for current player which are checkers.